[kernel] perf: Suppress AUX/OVERWRITE records
Message-id: <20190311094227.31342-2-jolsa@redhat.com>
Patchwork-id: 245735
O-Subject: [PATCH RHEL7 BZ1680074 001/170] perf: Suppress AUX/OVERWRITE records
Bugzilla: 1680074
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1680074
https://bugzilla.redhat.com/show_bug.cgi?id=1680074

upstream
========
commit 1627314fb54a33ebd23bd08f2e215eaed0f44712
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Wed Apr 4 17:53:23 2018 +0300

description
===========
It has been pointed out to me many times that it is useful to be able to
switch off AUX records to save the bandwidth for records that actually
matter, for example, in AUX overwrite mode.

The usefulness of PERF_RECORD_AUX is in some of its flags, like the
TRUNCATED flag that tells the decoder where exactly gaps in the trace
are.  The OVERWRITE flag, on the other hand will be set on every single
record in overwrite mode. However, a PERF_RECORD_AUX[flags=OVERWRITE] is
generated on every target task's sched_out, which over time adds up to a
lot of useless information.

If any folks out there have userspace that depends on a constant stream
of OVERWRITE records for a good reason, they'll have to let us know.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>