[tools] perf evsel: Mark a evsel as disabled when asking the kernel do disable it
Message-id: <20190311094227.31342-97-jolsa@redhat.com>
Patchwork-id: 245825
O-Subject: [PATCH RHEL7 BZ1680074 096/170] perf evsel: Mark a evsel as disabled when asking the kernel do disable it
Bugzilla: 1680074
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1680074
https://bugzilla.redhat.com/show_bug.cgi?id=1680074

upstream
========
commit b7e8452b860c299f342a012922bdd9ab8f2bb722
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Sat Oct 20 09:04:41 2018 -0300

description
===========
Because there may be more such events in the ring buffer that should be
discarded when an app decides to stop considering them.

At some point we'll do this with eBPF, this way we stop them at origin,
before they are placed in the ring buffer.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>