[tools] perf vendor events intel: Fix diverse typos
Message-id: <20190311094227.31342-160-jolsa@redhat.com>
Patchwork-id: 245889
O-Subject: [PATCH RHEL7 BZ1680074 159/170] perf vendor events intel: Fix diverse typos
Bugzilla: 1680074
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1680074
https://bugzilla.redhat.com/show_bug.cgi?id=1680074

upstream
========
commit b1d6f155e1bbb67778c17aba661fb4ea4e1a3641
Author: Ingo Molnar <mingo@kernel.org>
Date: Mon Dec 3 11:22:00 2018 +0100

description
===========
Go over the tools/ files that are maintained in Arnaldo's tree and
fix common typos: half of them were in comments, the other half
in JSON files.

( Care should be taken not to re-import these typos in the future,
  if the JSON files get updated by the vendor without fixing the typos. )

No change in functionality intended.

Committer notes:

This was split from a larger patch as there are code that is,
additionally, maintained outside the kernel tree, so to ease cherry
picking and/or backporting, split this into multiple patches.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>