[netdrv] ixgbe: fix warning: sysfs: cannot create duplicate filename
Message-id: <76dbec9c67f8dda13866c91ec9229ffe970172ea.1552577487.git.jkc@redhat.com>
Patchwork-id: 246141
O-Subject: [RHEL7.7 BZ 1688296 V2] ixgbe: Fix warning: sysfs: cannot create duplicate filename
Bugzilla: 1688296
RH-Acked-by: Stefano Brivio <sbrivio@redhat.com>
RH-Acked-by: Andrea Claudi <aclaudi@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1688296
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=20603644
Tested: boot/reboot tests on systems with multiple ixgbe adapters to ensure
	the error is not present
Upstream Status: RHEL only

Systems with more than 1 ixgbe network interface can try to create
multiple sysfs directories with the same name.  This results in the above
warning with a stack backtrace.  It can also cause a panic when the machine
is rebooted.

The problem is that in ixgbe_mii_bus_init() it is creating a string that is
used for the sysfs directory name but the string is truncated because
in RHEL7 the value for MII_BUS_ID_SIZE is 17 and in the upstream code
the value for MII_BUS_ID_SIZE is 61.

This commit uses pci device info to generate a unique string that
will not be truncated.

V2: Fix whitespace

Signed-off-by: Ken Cox <jkc@redhat.com>

Cc: Ken Cox <jkc@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>