[fs] ext4: try to initialize all groups we can in case of failure on ppc64
Message-id: <20190320142322.2848-2-lczerner@redhat.com>
Patchwork-id: 247113
O-Subject: [RHEL7.7 fs PATCH 1/2] ext4: try to initialize all groups we can in case of failure on ppc64
Bugzilla: 1079962
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

BZ 1079962
https://bugzilla.redhat.com/show_bug.cgi?id=1079962

BREW 20664047
https://brewweb.devel.redhat.com/taskinfo?taskID=20664047

commit bbdc322f2c600667e3d23dfadf8bbaad08f7edd3

Currently on the machines with page size > block size when initializing
block group buddy cache we initialize it for all the block group bitmaps
in the page. However in the case of read error, checksum error, or if
a single bitmap is in any way corrupted we would fail to initialize all
of the bitmaps. This is problematic because we will not have access to
the other allocation groups even though those might be perfectly fine
and usable.

Fix this by reading all the bitmaps instead of error out on the first
problem and simply skip the bitmaps which were either not read properly,
or are not valid.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>