[kernel] bpf: restrict unknown scalars of mixed signed bounds for unprivileged
Message-id: <20190328194906.20137-4-jolsa@redhat.com>
Patchwork-id: 248246
O-Subject: [PATCH RHEL7 BZ1673617 3/5] bpf: restrict unknown scalars of mixed signed bounds for unprivileged
Bugzilla: 1673617
CVE: CVE-2019-7308
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1673617
https://bugzilla.redhat.com/show_bug.cgi?id=1673617

upstream
========
commit 9d7eceede769f90b66cfa06ad5b357140d5141ed
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Thu Jan 3 00:58:32 2019 +0100

description
===========
For unknown scalars of mixed signed bounds, meaning their smin_value is
negative and their smax_value is positive, we need to reject arithmetic
with pointer to map value. For unprivileged the goal is to mask every
map pointer arithmetic and this cannot reliably be done when it is
unknown at verification time whether the scalar value is negative or
positive. Given this is a corner case, the likelihood of breaking should
be very small.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>