[lib] idr: free the top layer if idr tree has the maximum height
Message-id: <20190410124040.7070-1-vdronov@redhat.com>
Patchwork-id: 249528
O-Subject: [RHEL7.7 lib PATCH] idr: free the top layer if idr tree has the maximum height
Bugzilla: 1698115
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla: 1698115
Upstream Status: RHEL7-only, idr subsystem was rewritten from scratch in the upstream
Build info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=21043029
Test: a reproducer is in the Bugzilla ticket

There is a memory leak in the idr subsystem. The top layer is not freed when
idr with a tree of the maximum height of MAX_IDR_LEVEL is destroyed. This
happens because id becomes negative and a loop breaks before all ary[]
elements in the top layer (idp->top) are scanned. Fix this by explicit
check for an id overflow and a release of the top layer after the main loop.

Signed-off-by: Vladis Dronov <vdronov@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>