[redhat] kabi: adding symbol copy_user_generic_string (x86_64: arch/x86/kernel/x8664_ksyms_64.c)
Message-id: <20190411081226.32002-11-ckalina@redhat.com>
Patchwork-id: 249620
O-Subject: [RHEL7.7 PATCH v2 10/48] kabi: Adding symbol copy_user_generic_string (x86_64: arch/x86/kernel/x8664_ksyms_64.c)
Bugzilla: 1633921
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

Bugzilla: 1633921
https://bugzilla.redhat.com/show_bug.cgi?id=1633921

Description:
============
pdfs_ko: It is not used on the source of Primesoft Distributed File
System for Hadoop (P-DFS).

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

__attribute__ ( ( warn_unused_result ) ) unsigned long
    copy_user_generic_string ( void * , const void * , unsigned )

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: estmr Kalina <ckalina@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>