[fs] aio: get rid of unnecessary locking in aio_read_events_ring
Message-id: <x494l6npmx2.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 250903
O-Subject: [rhel7 patch] aio: get rid of unnecessary locking in aio_read_events_ring
Bugzilla: 1720750
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Ming Lei <ming.lei@redhat.com>

BZ: 1720750
Brew: 21277423
Upstream Status: RHEL-only

On certain hardware configurations, a CPU can be consumed by I/O
completion events leading to a softlockup.  Ming pointed out that the
aio completion lock may be contended, which could cause long stall times
in irq context.  Way, way, way back in the beginning of RHEL 7, we
pulled in patches that allowed the aio ring pages to be migrated.  Among
them was a fix for racing migratepage and event reaping.  That patch was
pulled in before it was cooked, and it added locking to
aio_read_events_ring.  Then, I later backported the fix that *actually*
went upstream, which was different from our prelimary version.  However,
I failed to undo the locking additions.  The result is we do extra
locking in the hot path that isn't needed.  We can safely remove that.

Ming tested this patch, and it resolves the issue.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>