[mm] oom_killer: Add task UID to printed info on an oom kill
Message-id: <1556127921-16584-1-git-send-email-jsavitz@redhat.com>
Patchwork-id: 250908
O-Subject: [RHEL7.7 v2] mm/oom_killer: Add task UID to printed info on an oom kill
Bugzilla: 1691184
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1691184
Upstream Status: RHEL-only, not upstream
Build info: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=21217160
Tested: Abused the RAM with simple program, UID was printed in dmesg

Modify oom killer message to print UID when a process is oom killed.

This is a forward port of a feature present in RHELs 5 & 6 in order
to meet a customer request:
	el5: e69cc61 [mm] oom killer output should display UID
	el6: 9021e2a [mm] oom killer output should display UID
The customer has requested that this feature be present in RHEL 7, and so
it shall be.

We plan to repost this patch to upstream with a more convincing
argument than last time.

Signed-off-by: Joel Savitz <jsavitz@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>