[x86] cpufeature: Add facility to check for min microcode revisions
Message-id: <20190429081454.24988-5-jolsa@redhat.com>
Patchwork-id: 251744
O-Subject: [PATCH RHEL7 BZ1482406 04/15] x86/cpufeature: Add facility to check for min microcode revisions
Bugzilla: 1482406
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1482406
https://bugzilla.redhat.com/show_bug.cgi?id=1482406

upstream
========
commit 0f42b790c9ba5ec2f25b7da8b0b6d361082d67b0
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Mon Feb 4 14:23:29 2019 -0800

description
===========
For bug workarounds or checks, it is useful to check for specific
microcode revisions.

Add a new generic function to match the CPU with stepping.
Add the other function to check the min microcode revisions for
the matched CPU.

A new table format is introduced to facilitate the quirk to
fill the related information.

This does not change the existing x86_cpu_id because it's an ABI
shared with modules, and also has quite different requirements,
as in no wildcards, but everything has to be matched exactly.

RHEL7
=====
Using x86_mask instead of x86_stepping which got renamed in upstream.

Signed-off-by: Jan Stancek <jstancek@redhat.com>