[x86] perf/x86: Support constraint ranges
Message-id: <20190429081454.24988-14-jolsa@redhat.com>
Patchwork-id: 251755
O-Subject: [PATCH RHEL7 BZ1482406 13/15] perf/x86: Support constraint ranges
Bugzilla: 1482406
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1482406
https://bugzilla.redhat.com/show_bug.cgi?id=1482406

upstream
========
commit 63b79f6ebc464afb730bc45762c820795e276da1
Author: Peter Zijlstra <peterz@infradead.org>
Date: Tue Apr 2 12:45:04 2019 -0700

description
===========
Icelake extended the general counters to 8, even when SMT is enabled.
However only a (large) subset of the events can be used on all 8
counters.

The events that can or cannot be used on all counters are organized
in ranges.

A lot of scheduler constraints are required to handle all this.

To avoid blowing up the tables add event code ranges to the constraint
tables, and a new inline function to match them.

Signed-off-by: Jan Stancek <jstancek@redhat.com>