netfilter: flow_table: add bpf hook on add/del actions
This work adds a programmable BPF hook to the add/del actions of the
nft flow table.  This allows using the nft path as a fallback for the
stack to do deep packet parsing, and allow an XDP fast-path program
to forward packets without needing to perform such deep inspection.
Alternatively, this allows for a tracing probe to be attached at the
flow-table points for additonal policy debugging.

Signed-off-by: Aaron Conole <aconole@bytheb.org>