[fs] ext4: check superblock mapped prior to committing
Message-id: <20190619140407.32403-1-lczerner@redhat.com>
Patchwork-id: 265651
O-Subject: [RHEL7.7 fs PATCH] ext4: check superblock mapped prior to committing
Bugzilla: 1721737
RH-Acked-by: Bill O'Donnell <billodo@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: Jon Derrick <jonathan.derrick@intel.com>

BZ 1721737
https://bugzilla.redhat.com/show_bug.cgi?id=1721737

BREW 22262836
https://brewweb.devel.redhat.com/taskinfo?taskID=22262836

commit a17712c8e4be4fa5404d20e9cd3b2b21eae7bc56

This patch attempts to close a hole leading to a BUG seen with hot
removals during writes [1].

A block device (NVME namespace in this test case) is formatted to EXT4
without partitions. It's mounted and write I/O is run to a file, then
the device is hot removed from the slot. The superblock attempts to be
written to the drive which is no longer present.

The typical chain of events leading to the BUG:
ext4_commit_super()
  __sync_dirty_buffer()
    submit_bh()
      submit_bh_wbc()
        BUG_ON(!buffer_mapped(bh));

This fix checks for the superblock's buffer head being mapped prior to
syncing.

[1] https://www.spinics.net/lists/linux-ext4/msg56527.html

Signed-off-by: Jon Derrick <jonathan.derrick@intel.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@kernel.org
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>