[char] random: move FIPS continuous test to output functions
Message-id: <20190628160503.smuyzl2wqwnbxold@gondor.apana.org.au>
Patchwork-id: 265940
O-Subject: [RHEL7.7 PATCH] random: move FIPS continuous test to output functions
Bugzilla: 1723910
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

BZ: 1723910
Brew: https//brewweb.engineering.redhat.com/brew/taskinfo?taskID=22418948
Upstream Status: Ignored/Rejected

For RHEL7 FIPS certification we need to certify DRBGs used in
user-space applications such as openssl.  These DRBGs are currently
seeded with /dev/urandom.  Previously /dev/urandom had a continuous
self-test that was added back in 2009.

The /dev/urandom self-test was effectively removed when the chacha20
RNG was introduced.  This patch (to reinstate it) was then submitted
but it was ignored.  The chacha20 RNG code was backported to RHEL7
in 2017.

The end result is that the FIPS certification of RHEL7.6/7.7 cannot
proceed.

We have a similar issue with RHEL8 but there we have the option of
introducing the self-test in user-space instead.  Apparently this
is not an option for RHEL7 for business reasons.  I am therefore
proposing that we apply this patch for RHEL7 and RHEL7 only, so
that RHEL7 FIPS certification can proceed.

Author: Stephan Mueller <smueller@chronox.de>

The current location of the FIPS continuous self test covers the
input_pool only. However, the FIPS continuous self test shall cover the
output of the random number generator, i.e. the blocking pool and the
ChaCha20 DRNG.

This patch therefore moves the continuous test to the output function
used for /dev/random. In addition, it adds the continuous test to the
ChaCha20 output function.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>