[mm] revert "don't split THP page when MADV_FREE syscall is called"
Message-id: <fbb18fcdccf482c83cc1eb8ca217668d29f35659.1561744567.git.aquini@redhat.com>
Patchwork-id: 265966
O-Subject: [RHEL7 BZ#1720278 PATCH v2 4/5] Revert "[mm] don't split THP page when MADV_FREE syscall is called"
Bugzilla: 1720278
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>

This reverts commit 104ce51651c28c83d315080ec810076230f4297b.

Back in rhel-7.6 development, we introduced a new behavior flag to madvise(2)
to allow pages to be released by applications and easily reclaimed by the
kernel, in case of memory pressure. MADV_FREE plays with clearing the pte
dirty bit in a given range to let the kernel know the pages do not need to
be re-sync'd to storage, thus flagging them as cheap reclaim candidates.

As a follow-up performance enhancement to MADV_FREE, it was also backported
a small set of patches to introduce madvise_free_huge_pmd() in order to
avoid splitting THPs when the syscall was called in THP aligned ranges.
The split would be deferred to the reclaim path, later, if the kernel
ended up deciding on reclaiming the pages within the marked range.

In order to get this THP splitting deferred, the original code
for __split_huge_page_map() was changed to remove its unconditional
call to pte_mkdirty() and the dirtiness handling was transferred to
the struct page. As dealing with THP splitting is racy, by nature,
the changes introduced for RHEL-7 backport "commit 104ce51651
([mm] don't split THP page when MADV_FREE syscall is called)" ended
up unveiling a bug in s390, which is an unique arch in respect to
dirty page detection, where regardless of utilizing MADV_FREE or
not, the new dirty bit handling would racily interfere with THP splitting
and with potential to cause data loss for the first page within the
compound block being splitted, in case of memory pressure reclaims.

IBM helped on isolating the issue and providing insights on how to fix it
without reverting the patches for the optimization aforementioned. I,
however, rather being safe than sorry, have decided to proceed with
the revert of the performance optimization for MADV_FREE in order to
get the code to a previous known stable state. Changes in the splitting
THP code have made the RHEL7 code substantially different from upstream
and having the code in an intermediate state might actually be source
of other subtle races hard to debug. Users of madvise(MADV_FREE) can
live with mandatory THP splitting at syscall time, so lets just revert
those bits.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>

Conflicts:
 mm/huge_memory.c
 * minor conflict due to RHEL-7 backport "d3b6a61266ef ([mm] thp: fix MADV_DONTNEED vs. MADV_FREE race)"

Signed-off-by: Rafael Aquini <aquini@redhat.com>