[cpuidle] cpuidle: dont call poll_idle_init() for every cpu
Message-id: <20190731132728.146526064@amt.cnet>
Patchwork-id: 267433
O-Subject: [RHEL 7.8 PATCH v3 01/17] cpuidle: dont call poll_idle_init() for every cpu
Bugzilla: 1734501
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

BZ: 1734501

Upstream status:
commit d7c7f103262bc2248548ed0e113e916e843c4eeb
of linux-2.6.git.

poll_idle_init() just initializes drv->states[0] and so that is
required to be done only once for each driver.  Currently, it is
called from cpuidle_enable_device() which is called for every CPU
that the driver supports.  That is not required, so move it to a
better place and call it from __cpuidle_register_driver() so that
the initialization is carried out only once.

Acked-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>