[cpuidle] cpuidle: poll_state: Revise loop termination condition
Message-id: <20190731132728.710302614@amt.cnet>
Patchwork-id: 267431
O-Subject: [RHEL 7.8 PATCH v3 08/17] cpuidle: poll_state: Revise loop termination condition
Bugzilla: 1734501
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

BZ: 1734501

Upstream status:
commit 01bad1c6896db021db82042e71c2bf1f97cc026b
of linux-2.6.git.

===

If need_resched() returns "false", breaking out of the loop in
poll_idle() will cause a new idle state to be selected, so in fact
it usually doesn't make sense to spin in it longer than the target
residency of the second state.  [Note that the "polling" state is
used only if there is at least one "real" state defined in addition
to it, so the second state is always there.]  On the other hand,
breaking out of it early (say in case the next state is disabled)
shouldn't hurt as it is polling anyway.

For this reason, make the loop in poll_idle() break if the CPU has
been spinning longer than the target residency of the second state
(the "polling" state can only be state[0]).

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>