[tools] perf bench numa: Fix cpu0 binding
Message-id: <20190808172320.7971-1-mpetlan@redhat.com>
Patchwork-id: 268622
O-Subject: [RHEL7.8 BZ1637662] perf bench numa: Fix cpu0 binding
Bugzilla: 1637662
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1637662
https://bugzilla.redhat.com/show_bug.cgi?id=1637662

Testing: Manually tested on a ppc64le machine with all CPUs on NUMA node 1.

upstream
========
commit 6bbfe4e602691b90ac866712bd4c43c51e546a60
Author: Jiri Olsa <jolsa@kernel.org>
Date:   Thu Aug 1 16:26:42 2019 +0200

description
===========
Michael reported an issue with perf bench numa failing with binding to
cpu0 with '-0' option.

  # perf bench numa mem -p 3 -t 1 -P 512 -s 100 -zZcm0 --thp 1 -M 1 -ddd
  # Running 'numa/mem' benchmark:

   # Running main, "perf bench numa numa-mem -p 3 -t 1 -P 512 -s 100 -zZcm0 --thp 1 -M 1 -ddd"
  binding to node 0, mask: 0000000000000001 => -1
  perf: bench/numa.c:356: bind_to_memnode: Assertion `!(ret)' failed.
  Aborted (core dumped)

This happens when the cpu0 is not part of node0, which is the benchmark
assumption and we can see that's not the case for some powerpc servers.

Using correct node for cpu0 binding.

Signed-off-by: Jan Stancek <jstancek@redhat.com>