[fs] nfs/nfsd/sunrpc: enforce transport requirements for NFSv4
Message-id: <20190809154204.25022-1-drosario@redhat.com>
Patchwork-id: 268703
O-Subject: [RHEL-7.8 BZ1733275 PATCH 3/4] nfs/nfsd/sunrpc: enforce transport requirements for NFSv4
Bugzilla: 1733275
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1733275

DESCRIPTION:
------------
NFSv4 requires a transport "that is specified to avoid network
congestion" (RFC 7530, section 3.1, paragraph 2).  In practical terms,
that means that you should not run NFSv4 over UDP. The server has never
enforced that requirement, however.

This patchset fixes this by adding a new flag to the svc_version that
states that it has these transport requirements. With that, we can check
that the transport has XPT_CONG_CTRL set before processing an RPC. If it
doesn't we reject it with RPC_PROG_MISMATCH.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>

UPSTREAM STATUS:
----------------
commit 5283b03ee5cd28d516646298bead09b238d92ddc

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?id=5283b03ee5cd28d516646298bead09b238d92ddc
Signed-off-by: Jan Stancek <jstancek@redhat.com>