[s390] s390: use S390_lowcore instead of tmp variable to load cr7 in set_fs
Message-id: <20190822131952.26492-1-yauheni.kaliuta@redhat.com>
Patchwork-id: 269489
O-Subject: [RHEL7 PATCH v2] s390: use S390_lowcore instead of tmp variable to load cr7 in set_fs
Bugzilla: 1724027
RH-Acked-by: Philipp Rudo <prudo@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Cornelia Huck <cohuck@redhat.com>

Bugzilla: https://bugzilla.redhat.com/1724027
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=23157041
Upstream: RHEL-only
Testing: test_verifier
         /usr/share/bcc/tools/dcsnoop

__ctl_load(__pto, 7, 7), which is 'lctlg' instruction, requires the
argument to be doubleword (8 bytes) aligned:

```
The second operand must be designated on a word boundary for LCTL or
on a doubleword boundary for LCTLG; otherwise, a specification
exception is recognized.
```

With the current implementation the value to load is prepared in a
local variable (allocated on stack) which not necessary properly
aligned, what causes crashes like in the case of the bug when
running specific BPF program.

Fix it by using the S390_lowcore.kernel_asce or
S390_lowcore.user_asce directly, instead of local variable, in the
condition branches.

The problem fixed upstream with the patch:

b5a882fcf146 s390: restore address space when returning to user space

but it adresses another more broad issue, so this patch is not
considered as the backport. The hunk from the upstream patch is
changed (removed setting of CPU flag) to make it compilable and
avoid bringing extra functionality change.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>