[sound] ALSA: line6: Fix write on zero-sized buffer
Message-id: <1567579972674204-148-patchset-jkysela@redhat.com>
Patchwork-id: 270799
O-Subject: [RHEL 7.8 PATCH 147/186] ALSA: line6: Fix write on zero-sized buffer
Bugzilla: 1733130
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

LINE6 drivers allocate the buffers based on the value returned from
usb_maxpacket() calls.  The manipulated device may return zero for
this, and this results in the kmalloc() with zero size (and it may
succeed) while the other part of the driver code writes the packet
data with the fixed size -- which eventually overwrites.

This patch adds a simple sanity check for the invalid buffer size for
avoiding that problem.

Reported-by: syzbot+219f00fb49874dcaea17@syzkaller.appspotmail.com
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 3450121997ce872eb7f1248417225827ea249710)
Bugzilla: 1733130
Signed-off-by: Jan Stancek <jstancek@redhat.com>