[sound] ALSA: usb-audio: Skip bSynchAddress endpoint check if it is invalid
Message-id: <1567579972674204-158-patchset-jkysela@redhat.com>
Patchwork-id: 270822
O-Subject: [RHEL 7.8 PATCH 157/186] ALSA: usb-audio: Skip bSynchAddress endpoint check if it is invalid
Bugzilla: 1733130
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The Linux kernel assumes that get_endpoint(alts,0) and
get_endpoint(alts,1) are eachothers feedback endpoints.
To reassure that validity it will test bsynchaddress to comply with that
assumption. But if the bsyncaddress is 0 (invalid), it will flag that as
a wrong assumption and return an error.
Fix: Skip the test if bSynchAddress is 0.
Note: those with a valid bSynchAddress should have a code quirck added.

Signed-off-by: Ard van Breemen <ard@kwaak.net>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 1b34121d9f26d272b0b2334209af6b6fc82d4bf1)
Bugzilla: 1733130
Signed-off-by: Jan Stancek <jstancek@redhat.com>