[md] md: add block tracing for bio_remapping
Message-id: <1567660618-16591-3-git-send-email-xni@redhat.com>
Patchwork-id: 270930
O-Subject: [RHEL7.8 PATCH V2 2/5] md: add block tracing for bio_remapping
Bugzilla: 1741466
RH-Acked-by: Nigel Croxon <ncroxon@redhat.com>
RH-Acked-by: Ming Lei <ming.lei@redhat.com>

The block tracing infrastructure (accessed with blktrace/blkparse)
supports the tracing of mapping bios from one device to another.
This is currently used when a bio in a partition is mapped to the
whole device, when bios are mapped by dm, and for mapping in md/raid5.
Other md personalities do not include this tracing yet, so add it.

When a read-error is detected we redirect the request to a different device.
This could justifiably be seen as a new mapping for the originial bio,
or a secondary mapping for the bio that errors.  This patch uses
the second option.

When md is used under dm-raid, the mappings are not traced as we do
not have access to the block device number of the parent.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 109e37653033a5fcd3bf8cab0ed6a7ff433f758a)
Signed-off-by: Xiao Ni <xni@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>

Conflicts:
    drivers/md/linear.c
    drivers/md/raid0.c
    drivers/md/raid1.c
    drivers/md/raid10.c
linear.c has been removed.
For raid0.c rhel7 uses bio_pair, upstream uses bio_split and bio_chain in
raid0_make_request
For raid1, rhel7 already backported 3b046a9(Refactor raid1_make_request)
For raid10, rhel7 already backported 24f0bee(Refactor raid10_make_request)
For this patch, it just needs some remap events. It doesn't change the I/O
logic.