[fs] ext4: cleanup bh release code in ext4_ind_remove_space()
Message-id: <20190916064842.2680-12-lczerner@redhat.com>
Patchwork-id: 271903
O-Subject: [RHEL7.8 fs PATCH v2 11/17] ext4: cleanup bh release code in ext4_ind_remove_space()
Bugzilla: 1747387
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: "zhangyi (F)" <yi.zhang@huawei.com>

BZ 1747387
https://bugzilla.redhat.com/show_bug.cgi?id=1747387

BREW 23490730
https://brewweb.devel.redhat.com/taskinfo?taskID=23490730

commit 5e86bdda41534e17621d5a071b294943cae4376e

Currently, we are releasing the indirect buffer where we are done with
it in ext4_ind_remove_space(), so we can see the brelse() and
BUFFER_TRACE() everywhere.  It seems fragile and hard to read, and we
may probably forget to release the buffer some day.  This patch cleans
up the code by putting of the code which releases the buffers to the
end of the function.

Signed-off-by: zhangyi (F) <yi.zhang@huawei.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Jan Kara <jack@suse.cz>
Signed-off-by: Jan Stancek <jstancek@redhat.com>