[netdrv] ibmvnic: Do not process reset during or after device removal
Message-id: <20190918142240.938.16086.sendpatchset@stevebest>
Patchwork-id: 272408
O-Subject: [PATCH RHEL7.8 BZ1751815 1/3] ibmvnic: Do not process reset during or after device removal
Bugzilla: 1751815
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jerry Snitselaar <jsnitsel@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1751815

Description:
------------
Currently, the ibmvnic driver will not schedule device resets
if the device is being removed, but does not check the device
state before the reset is actually processed. This leads to a race
where a reset is scheduled with a valid device state but is
processed after the driver has been removed, resulting in an oops.

Fix this by checking the device state before processing a queued
reset event.

Reported-by: Abdul Haleem <abdhalee@linux.vnet.ibm.com>
Tested-by: Abdul Haleem <abdhalee@linux.vnet.ibm.com>
Signed-off-by: Thomas Falcon <tlfalcon@linux.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit 36f1031c51a2538e5558fb44c6d6b88f98d3c0f2

https://git.kernel.org/pub/scm/linux/kernel/git/davem/net.git/commit/?id=36f1031c51a2538e5558fb44c6d6b88f98d3c0f2

----------------
Steve Best

Signed-off-by: Jan Stancek <jstancek@redhat.com>