[fs] revert "[fs] nfsd: Implement the COPY call"
Message-id: <20190924215402.7282-2-bfields@redhat.com>
Patchwork-id: 274524
O-Subject: [Patch RHEL7 fs 2/2] Revert "[fs] nfsd: Implement the COPY call"
Bugzilla: 1688067
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1688067
Build Info: https://brewweb.devel.redhat.com/taskinfo?taskID=23692234
Upstream Status: RHEL-only

This reverts commit f84cefe56c9ec64e384834aba134c3bfb1042e4a.

RHEL7 has no nfs-exportable filesystems with ->copy_file_range.  The
only reason the NFS COPY operation was useful on our server was because
of the fallback to splice--but we're removing that.  The only risk here
is to 3rd-party filesystems that want to support NFS COPY.  I'm not
aware of any.

We could probably remove support for CLONE as well, as far as I know it
is similarly not supported anywhere.  But I'm also not aware of any
issue with it, so I'm leaving it be for now.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>