[tools] perf tools: Add a 'percore' event qualifier
Message-id: <20191024105146.30745-9-jolsa@redhat.com>
Patchwork-id: 281070
O-Subject: [PATCH RHEL7.8 BZ1640900 08/13] perf tools: Add a 'percore' event qualifier
Bugzilla: 1640900
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1640900
https://bugzilla.redhat.com/show_bug.cgi?id=1640900

upstream
========
commit 064b4e82aa1633c27c383cc686b87ced57e072d1
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Fri Apr 12 21:59:47 2019 +0800

description
===========
Add a 'percore' event qualifier, like cpu/event=0,umask=0x3,percore=1/,
that sums up the event counts for both hardware threads in a core.

We can already do this with --per-core, but it's often useful to do
this together with other metrics that are collected per hardware thread.
So we need to support this per-core counting on a event level.

This can be implemented in only the user tool, no kernel support needed.

 v4:
 ---
 1. Add Arnaldo's patch which updates the documentation for
    this new qualifier.
 2. Rebase to latest perf/core branch

 v3:
 ---
 Simplify the code according to Jiri's comments.
 Before:
   "return term->val.percore ? true : false;"
 Now:
   "return term->val.percore;"

 v2:
 ---
 Change the qualifier name from 'coresum' to 'percore' according to
 comments from Jiri and Andi.

Signed-off-by: Jan Stancek <jstancek@redhat.com>