[md] md: support for queue flag QUEUE_FLAG_NO_SG_MERGE
Message-id: <20191104143304.29944-1-ncroxon@redhat.com>
Patchwork-id: 282487
O-Subject: [RHEL7 PATCH 1/1] md: support for queue flag QUEUE_FLAG_NO_SG_MERGE
Bugzilla: 1767472
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>

Brew: 24387948
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=24387948

Bugzilla:  1767472
https://bugzilla.redhat.com/show_bug.cgi?id=1767472

Upstream: No
RHEL7 ONLY

Description:
All RAID drivers do not set the queue flag of QUEUE_FLAG_NO_SG_MERGE
if any of underlying queues set this flag. Upstream removed this
functionality but it was to invasive for RHEL7. So this is the fix
for MD.

This change is do to the following DM's commit for handling the issue:
      ad43261404c0 [md] dm-table: propagate QUEUE_FLAG_NO_SG_MERGE

Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>